renaming DeleteHandler to CleanupHandler
