+++ src/protagonist/CleanupHandler/Properties/launchSettings.json
