+++ src/protagonist/CleanupHandler/appsettings.json
@@ -1,0 +1,22 @@
+﻿{
+  "Serilog": {
+    "Using": [
+      "Serilog.Sinks.Console"
+    ],
+    "MinimumLevel": {
+      "Default": "Debug",
+      "Override": {
+        "Microsoft": "Warning",
+        "System": "Warning"
+      }
+    },
+    "WriteTo": [
+      {
+        "Name": "Console"
+      }
+    ],
+    "Properties": {
+      "ApplicationName": "CleanupHandler"
+    }
+  }
+}
