--- src/protagonist/DeleteHandler/appsettings.json
@@ -1,22 +1,0 @@
-﻿{
-  "Serilog": {
-    "Using": [
-      "Serilog.Sinks.Console"
-    ],
-    "MinimumLevel": {
-      "Default": "Debug",
-      "Override": {
-        "Microsoft": "Warning",
-        "System": "Warning"
-      }
-    },
-    "WriteTo": [
-      {
-        "Name": "Console"
-      }
-    ],
-    "Properties": {
-      "ApplicationName": "DeleteHandler"
-    }
-  }
-}
