+++ src/protagonist/protagonist.sln
@@ -65,7 +65,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Engine", "Engine\Engine.csp
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Engine.Tests", "Engine.Tests\Engine.Tests.csproj", "{EF411F9A-C628-4786-92D4-D40A3EDE7222}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeleteHandler", "DeleteHandler\DeleteHandler.csproj", "{A8760A1D-8E1B-4822-954C-41188BD9745F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CleanupHandler", "CleanupHandler\CleanupHandler.csproj", "{A8760A1D-8E1B-4822-954C-41188BD9745F}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeleteHandlerTests", "DeleteHandlerTests\DeleteHandlerTests.csproj", "{E0663AC2-FCCC-4660-999F-53AF37F8C63F}"
 EndProject
